Create first web page Modify index.html by adding link to first web page Rename pages/media to recipes-rest/images
The first web page representing first recipe is added initally only
including title of dish and images.
Then in order to make it acessable as one of the website recipes
home page is modified to include link element enabeling acess to it
as part of website.
Finally pages/media is renimed to recipes-rest/images due to latter
being more descriptive of actual content present in directory
due to it for now only contaning images